+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# 2021-06-26
+
+* Add option do_logNormCounts, and try to not error-out on completely empty cells.
+
 # 2021-06-14
 
 * Default to unstranded gaps (i.e. don't extend into gene on opposite strand).